Add unit tests for the contact support prompt
